% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/scklass_dist.R
\name{calcDistances}
\alias{calcDistances}
\title{Distance calculation}
\usage{
calcDistances(train_data, test_data, p = 2)
}
\arguments{
\item{train_data}{A data frame or a matrix where rows are observations and columns are features. This is the dataset to which distances are calculated (the training set)}

\item{test_data}{A data frame or a matrix where rows are observations and columns are features. This is the dataset from which distances are calculated (the test set)}

\item{p}{The type of distance being calculated (e.g. \code{p=2} corresponds to the Euclidean distance)}
}
\value{
A matrix where rows are the observations in the 'From' dataset and columns are observations in the 'To' dataset and cells contain the ordering (position) of neighbours for each of the observations in the 'From' dataset.
}
\description{
Calculates the distances from each observation in a data set (for example the test set) to every data point in a dataset (for example the training set). The two datasets could be the same when no test set is available.
}
\examples{
# create artificial dataset
train_data <- matrix(rnorm(200), ncol=2)
test_data  <- matrix(rnorm(50), ncol=2)
# calculate the Euclidean distances from each observation in \\code{test_data} to every observation in \\code{train_data}
calcDistances(train_data, test_data, p=2)
}

