% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/scklass_knn.R
\name{knnClassifier}
\alias{knnClassifier}
\title{k-NN Classification}
\usage{
knnClassifier(true_labels, obs, neighbours, labels, k = 1)
}
\arguments{
\item{true_labels}{A vector, of the same length as the number of rows in the training dataset containing the true label of each observation in the training dataset}

\item{obs}{The number of observations to classify}

\item{neighbours}{A matrix of distances between the test set and the training set obtained with the \code{calcDistances} function}

\item{labels}{A vector of unique labels/classes (needs to include all the unique labels making up the \code{true_labels} vector)}

\item{k}{The number of neighbours to consider in the classification}
}
\value{
A vector with the predicted label for each observation in the test dataset
}
\description{
Carries out a k-NN classification
}
\examples{
# create artificial dataset
train_data <- matrix(rnorm(200), ncol=2)
test_data  <- matrix(rnorm(50), ncol=2)
# create labels
labels <- c(1,2,3)
true_labels <- c(rep(labels[1],33), rep(labels[2],33), rep(labels[3],34))
# calculate distances
neighbours <- calcDistances(train_data, test_data, p=2)
# define the number of test observations to consider
obs <- nrow(test_data)
# carry out the k-NN classification
knnClassifier(true_labels,obs,neighbours,labels,k=3)
}

